Replace submodules with vcpkg

+++ Tests/ManualTests/APICSolver2Tests.cpp
@@ -305,7 +305,7 @@ CUBBYFLOW_BEGIN_TEST_F(APICSolver2, Circular)
     std::mt19937 rng;
     std::uniform_real_distribution<> dist(0, 1);
 
-    for (int i = 0; i < 4 * resX * resX; ++i)
+    for (size_t i = 0; i < 4 * resX * resX; ++i)
     {
         Vector2D pt{ dist(rng), dist(rng) };
 
@@ -352,7 +352,7 @@ CUBBYFLOW_BEGIN_TEST_F(APICSolver2, CircularWithFriction)
     std::mt19937 rng;
     std::uniform_real_distribution<> dist(0, 1);
 
-    for (int i = 0; i < 4 * resX * resX; ++i)
+    for (size_t i = 0; i < 4 * resX * resX; ++i)
     {
         Vector2D pt{ dist(rng), dist(rng) };
